requirements.txt
# Python 3.8.10

aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
django-environ==0.9.0
django-ninja==0.19.1
frozenlist==1.3.1
h11==0.13.0
idna==3.3
multidict==6.0.2
pycodestyle==2.9.0
pydantic==1.9.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
yarl==1.8.1
